Change license
This license is better because it covers contributions to the code from contributors. The BSD license does not cover that and so it's not clear what status PR's from others would be. With this, we can be sure that submissions are available for us to use and won't be yanked back in the future.