Adds a tabbed panel API - untested
This commit is mid-conversion of the settings panel and flags to use the
tabbed panel. There are a few issues that need to be addressed. One is
that flags can be asked to create their own panel item based on a user
and world. This is useful for the click listeners because when a flag is
clicked, it can update itself in a panel. If flags do not do this, then
a settings panel will need to keep track on where flags are.

Anyway, more work to do but I'm sharing the progress so far.