Adds admin command to change a player's settings.
Requires addon to add the admin settings command.

https://github.com/BentoBoxWorld/BentoBox/issues/59

This also changes TabbedPanel to extend Panel so that when an icon is
clicked, the context can be gathered by the click listener via the
getActiveTab method. In this case, the CycleClick and IslandToggleClick
classes cast the tab to the SettingsTab so they can retrieve the island
that the panel is referring too. This is required in the case where an
admin is setting a user's island settings. Previously the context of a
panel was only every about the user who opened the panel.