Enables DataObjects to define their own table names.
In BentoBox the implicit naming strategy for SQL table names and Mongo
collections is the canonical name of the class. This change enables
authors to change this behavior by implementing a physical naming
strategy via the @Table type annotation on DataObject. This approach
mirrors systems such as Hibernate.

*Automatic migration*
If a DataObject uses the @Table annotation and the database has
canonical class named tables, then the old table will renamed to the new
table automatically.

With shorter table names, the limitation on the table name prefix has
been lifted.