Switch to island homes from player homes.
Stores home locations and max homes in the Island object. Adds commands
required to manage home names, specifically rename and delete. I did not
add list as there is tab complete on island go, but it may be required.


To Do:

1. Test
2. Add max homes setting when owner logs in.
3. Fix unit tests - they are probably broken
4. Add a conversion from player home locations to the island homes. Not
sure what the best way to do this is. Use the owner's?
5. Localization of new strings