Revert 6e734fc3432b1d435d1c02a462021d25195ec4af
The commit was wrong. Protection range should not be decreased, as that would mean that max X and Z blocks are not included in island protection bounding box.
This change is the same as 3c65194dfb332ac6beaf8d956bbf4350c0cd9fa5 but was missed when it was fixed.

Relate to #473