Address #2769 with both refactoring and defensive code
This uses the portal cooldown function to debounce the events. The
entities immediately enter the portal after they exit because they tend
not to move if there is no player on the other side, however they do not
teleport, so they don't exit.

In addition, the inPortal and inTeleport sets now use ExpiringSet
defensively to remove any values after 10 seconds, which should be more
than enough of time to complete any teleporting. This is defensive code
but will prevent memory leak build up.