Pops the player to the nether an back when they make an island to force
blocks to be loaded. Pasting blocks using NMS is fast, but doesn't
update the client. If the client is teleporting from another world, then
there is no issue, but if they are in the same world and close to the
new island spot, they'll float in nothing unless they log out and back
in.

I tried using NMS to force chunk reloads, but it didn't work and it was
complicated. A reliable approach is to teleport the player to another
world and back immediately. This works, but requires another world. I
picked the nether of the game world because it is usually there. If it
isn't then this doesn't work, sorry.